+++ src/main/java/com/pepero/jcb/api/GaviotaAnalyzer.java
@@ -2,9 +2,8 @@ package com.pepero.jcb.api;
 
 import com.pepero.jcb.api.dto.GaviotaMoveDTO;
 import com.pepero.jcb.api.dto.MoveInfo;
-import com.pepero.jcb.api.exception.VariantNotMatchException;
+import com.pepero.jcb.api.exception.game.VariantNotMatchException;
 import com.pepero.jcb.api.gaviota.GaviotaTablebase;
-import com.pepero.jcb.core.ChessboardUtils;
 import com.pepero.jcb.core.constant.MoveCache;
 import com.pepero.jcb.core.Chessboard;
 import com.pepero.jcb.core.GameVariant;