+++ src/main/java/com/pepero/jcb/api/arena/ProcessUCIEngineFactory.java
@@ -1,11 +1,10 @@
 package com.pepero.jcb.api.arena;
 
-import com.pepero.jcb.api.exception.UCIEngineException;
+import com.pepero.jcb.api.exception.engine.UCIEngineException;
 import com.pepero.jcb.api.uci.UCIEngineWrapper;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;