+++ src/main/java/com/pepero/jcb/api/exception/engine/EngineArenaException.java
@@ -1,4 +1,4 @@
-package com.pepero.jcb.api.exception.arena;
+package com.pepero.jcb.api.exception.engine;
 
 public class EngineArenaException extends RuntimeException {
     public EngineArenaException(String message) {