+++ src/main/java/com/pepero/jcb/api/exception/game/HistoryTreeException.java
@@ -1,4 +1,4 @@
-package com.pepero.jcb.api.exception;
+package com.pepero.jcb.api.exception.game;
 
 /**
  * Thrown by the history tree's root node deleted, etc.