+++ src/main/java/com/pepero/jcb/api/exception/game/MoveNotFoundException.java
@@ -1,4 +1,4 @@
-package com.pepero.jcb.api.exception;
+package com.pepero.jcb.api.exception.game;
 
 public class MoveNotFoundException extends RuntimeException {
     public MoveNotFoundException(String message) {