+++ src/test/java/com/pepero/jcb/api/gaviota/GaviotaTest.java
@@ -4,7 +4,7 @@ import com.pepero.jcb.api.ChessGame;
 import com.pepero.jcb.api.GaviotaAnalyzer;
 import com.pepero.jcb.api.dto.MoveInfo;
 import com.pepero.jcb.api.enums.GameOverReason;
-import com.pepero.jcb.api.exception.IllegalMoveException;
+import com.pepero.jcb.api.exception.game.IllegalMoveException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;