+++ src/test/java/com/pepero/jcb/example/VariantGameExample.java
@@ -4,7 +4,7 @@ import com.pepero.jcb.api.ChessGame;
 import com.pepero.jcb.api.enums.GameOverReason;
 import com.pepero.jcb.api.enums.PieceType;
 import com.pepero.jcb.api.enums.Square;
-import com.pepero.jcb.api.exception.IllegalMoveException;
+import com.pepero.jcb.api.exception.game.IllegalMoveException;
 import com.pepero.jcb.core.GameVariant;
 
 import java.io.BufferedReader;